+++ v2/db/shared.js
@@ -1,0 +1,11 @@
+// Used to prevent duplicate entries for an attribute
+function guardItem(prefix, value) {
+  return {
+    TableName: MAIN_TABLE_NAME,
+    ConditionExpression: 'attribute_not_exists(pk)',
+    Item: {
+      pk: `${prefix}#${value}`,
+      sk: value,
+    },
+  };
+}
